refactor: group bridging route contents
